requirements.txt
beautifulsoup4==4.13.3
faiss-cpu==1.10.0
matplotlib==3.9.2
numpy==2.1.3
openai==1.61.1
pandas==2.2.3
python-dotenv==1.0.1
seaborn==0.13.2
tqdm==4.67.1
tiktoken==0.9.0